// eshop-web/src/main/resources/client/src/app/analytic-logging/enums/basket-item-source-desc.enum.ts
export enum BasketItemSourceDesc {
    VIN_SEARCH = 'vin-search',
    VEHICLE_CODE_SEARCH = 'vehicle-code-search',
    VEHICLE_DESC_SEARCH = 'vehicle-desc-search',
    FULL_TEXT_SEARCH = 'full-text-search',
    FAST_SCAN = 'fast-scan',
    ARTICLE_SEARCH = 'article-search',
    ARTICLE_DEEP_LINK = 'article-deep-link',
    SHOPPING_LIST_SEARCH = 'shopping-list-search',
    WSP_CATALOG = 'wsp-catalog',
    TYRE_SEARCH = 'tyre-search',
    BATTERY_SEARCH = 'battery-search',
    BULB_SEARCH = 'bulb-search',
    OIL_SEARCH = 'oil-search',
    THULE_CATALOG = 'thule-catalog',
    DVSE_CATALOG = 'dvse-catalog',
    ORDER_HISTORY = 'order-history',
    INVOICE_ARCHIVE = 'invoice-archive',
    WSS_FC_ORDER = 'wss-fc-order',
    WSS_FC_SHOPPING_LIST_SEARCH = 'wss-fc-shopping-list-search',
    VEHICLE_MAKEMODELTYPE_SEARCH = 'vehicle-makemodeltype-search',
    VEHICLE_MAKEMODELTYPE_SEARCH_PAGED = 'vehicle-makemodeltype-search-paged',
    VEHICLE_FULL_TEXT_SEARCH = 'vehicle-full-text-search',
    VEHICLE_ID_SEARCH = 'vehicle-id-search',
    VEHICLE_MOTO_MAKEMODEL_SEARCH = 'vehicle-moto-makemodel-search',
    VEHICLE_FAVORITE = 'vehicle-favorite',
    ARTICLE_FAVORITE = 'article-favorite'
}